minor corrections: double

+++ src/DeclMap.h
@@ -12,17 +12,17 @@ class DeclMap
     public:
         static DeclMap& Instance();
 
-        int getValue(const string & ident);
+        double getValue(const string & ident);
         bool getIsConst(const string & ident);
         bool getIsInitialized(const string & ident);
         bool getVarIsReaded(const string & ident);
 
         void clearMap() {m_declMap.clear();};
-        void setValue(const string & ident, int val);
+        void setValue(const string & ident, double val);
         void setUsed(const string & ident);
         void setInitialized(const string & ident);
         void setVarIsReaded(const string & ident);
-        void addIdent(const string & ident, bool isConst, int val = 0);
+        void addIdent(const string & ident, bool isConst, double val = 0);
         bool checkIdent(const string & ident);
         void print();
         void staticCheck();
@@ -37,7 +37,7 @@ class DeclMap
         struct IdentValue
         {
             bool isConst, isInitialized, isUsed, varIsReaded;
-            int value;
+            double value;
         };
 
         map<string,IdentValue> m_declMap;
